pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "common_helper"   # must be unique on PyPI
version = "0.1.0"
description = "Reusable helpers for OTP, JWT authentication, and JSON serialization"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Your Name", email = "you@example.com" }
]
dependencies = [
    "pyjwt>=2.0.0",
    "fastapi>=0.95.0",
    "requests>=2.25.0"
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/yourusername/myhelpers"
